In the description you can see what I did
So the reason why my code was screwed up was because recursiveOrIterative was not an array. So I switched it to an array and nothing changed but then I thought to myself, I need to defrefrence it in my if else statement below. So i did that and itr worked. Also by me using r and i, i was doing other commands in the command line so I changed it to L and T.