So my code works and Ill explain why
So when I say _argc == 4, this means that I am expecting 3 actual arguments. The reason why is because the first index is meant for the executable. This is why in test.sh, we have ./fib 5 r test1.txt. ./fib is the executable so it needs to be first.

Now the reason why i did argv[3] is because the 3rd input is file name. One thing I was getting confused about is that if argv seperates each letter into a cell. The answer is no and the reason why is because it separates by white space and not by each individual cell.

Now i did atoi(__argv[1]) because I need to convert the ascii number to integer.

I did *argv[2] == 'i' because I need to see if its recursive or iterative.